I think I'm done??? Still a success!
I managed to get all of the loops to nest into the right place.  But I'm
not 100% sure whether I fulfilled the request of the assignment, on a
number of levels.  But still, I declare this code a raving success! This
entire thing came out of my own brain (aided by google) with ZERO imput
from Jim: he's in Denver! Plus it only took a few days of active work. :-D